feat: add main menu launcher and restructure app navigation
