+++ gui/__init__.py
@@ -1,2 +1,3 @@
+from .new_simulation_menu import NewSimulationWindow
 from .main_menu import MainMenuWindow
 from .setup_loop import SetupLoop
